remove ProtonVPN server file

+++ tests/test_providers.py
@@ -9,7 +9,8 @@ from ensure_vpn.constants import PROTONVPN_SERVER_FILE_PATH
 providers = ("mullvad", "nordvpn", "protonvpn")
 custom_ips = ("217.138.222.100", "217.138.222.0/24")
 
-os.remove(PROTONVPN_SERVER_FILE_PATH)
+if os.path.isfile(PROTONVPN_SERVER_FILE_PATH):
+    os.remove(PROTONVPN_SERVER_FILE_PATH)
 
 @pytest.mark.vcr(record_mode="none")
 @pytest.mark.parametrize("ip_or_provider", providers + custom_ips)
